refactor(exception): replace service runtime exceptions with custom exceptions

+++ backend/src/main/java/org/travel_stories/service/FollowService.java
@@ -6,6 +6,9 @@ import org.springframework.transaction.annotation.Transactional;
 import org.travel_stories.dto.FollowResponseDto;
 import org.travel_stories.entity.Follow;
 import org.travel_stories.entity.User;
+import org.travel_stories.exception.InvalidOperationException;
+import org.travel_stories.exception.ResourceAlreadyExistsException;
+import org.travel_stories.exception.ResourceNotFoundException;
 import org.travel_stories.repository.FollowRepository;
 import org.travel_stories.repository.UserRepository;
 
@@ -20,31 +23,41 @@ public class FollowService {
     private final FollowRepository followRepository;
     private final UserRepository userRepository;
 
-    public String follow(UUID followerId, UUID followingId){
-        if (followerId.equals(followingId)){
-            throw new IllegalArgumentException("User cannot follow themself.");
+    @Transactional
+    public String follow(UUID followerId, UUID followingId) {
+
+        if (followerId.equals(followingId)) {
+            throw new InvalidOperationException("User cannot follow themselves.");
         }
 
-        if (followRepository.existsByFollowerUserIdAndFollowingUserId(followerId, followingId)){
-            return "Exists.";
+        if (followRepository.existsByFollowerUserIdAndFollowingUserId(followerId, followingId)) {
+            throw new ResourceAlreadyExistsException("User is already being followed.");
         }
 
         User follower = userRepository.findById(followerId)
-                .orElseThrow(() -> new RuntimeException("Follower not found."));
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Follower not found."));
 
         User following = userRepository.findById(followingId)
-                .orElseThrow(() -> new RuntimeException("Following user not found."));
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Following user not found."));
 
         Follow follow = new Follow();
-
         follow.setFollower(follower);
         follow.setFollowing(following);
 
         followRepository.save(follow);
+
         return "Followed.";
     }
 
-    public void unfollow(UUID followerId, UUID followingId){
+    @Transactional
+    public void unfollow(UUID followerId, UUID followingId) {
+
+        if (!followRepository.existsByFollowerUserIdAndFollowingUserId(followerId, followingId)) {
+            throw new ResourceNotFoundException("Follow relationship not found.");
+        }
+
         followRepository.deleteByFollowerUserIdAndFollowingUserId(followerId, followingId);
     }
